add zlib license to header and source files
while this code is original, the zlib algorithm was studied
to create an English language description. in any, case this
change is to credit Mark Adler and Jean-loup Gailly for the
work on the Rabin-Karp algorithm in the zlib compressor, and
to update the license so that it is zlib compatible.